fix storybook demos and address PR feedback from Erik
