+++ src/realm-roles/RealmRoleForm.tsx
@@ -29,18 +29,18 @@ import { useConfirmDialog } from "../components/confirm-dialog/ConfirmDialog";
 import { RoleAttributes } from "./RoleAttributes";
 
 type RoleFormType = {
-  form: UseFormMethods;
-  save: SubmitHandler<RoleRepresentation>;
-  editMode: boolean;
+  form?: UseFormMethods;
+  save?: SubmitHandler<RoleRepresentation>;
+  editMode?: boolean;
 };
 
-const RoleForm = ({ form, save, editMode }: RoleFormType) => {
+export const RoleForm = ({ form, save, editMode }: RoleFormType) => {
   const { t } = useTranslation("roles");
   const history = useHistory();
   return (
     <FormAccess
       isHorizontal
-      onSubmit={form.handleSubmit(save)}
+      onSubmit={form!.handleSubmit(save!)}
       role="manage-realm"
       className="pf-u-mt-lg"
     >
@@ -48,11 +48,11 @@ const RoleForm = ({ form, save, editMode }: RoleFormType) => {
         label={t("roleName")}
         fieldId="kc-name"
         isRequired
-        validated={form.errors.name ? "error" : "default"}
+        validated={form!.errors.name ? "error" : "default"}
         helperTextInvalid={t("common:required")}
       >
         <TextInput
-          ref={form.register({ required: true })}
+          ref={form!.register({ required: true })}
           type="text"
           id="kc-name"
           name="name"
@@ -71,7 +71,7 @@ const RoleForm = ({ form, save, editMode }: RoleFormType) => {
       >
         <TextArea
           name="description"
-          ref={form.register({
+          ref={form!.register({
             maxLength: {
               value: 255,
               message: t("common:maxLength", { length: 255 }),
@@ -79,7 +79,7 @@ const RoleForm = ({ form, save, editMode }: RoleFormType) => {
           })}
           type="text"
           validated={
-            form.errors.description
+            form!.errors.description
               ? ValidatedOptions.error
               : ValidatedOptions.default
           }
